Database Gender Refactor
Added @map for gender enum. Postgres does that support directly changing ENUM to String. So Gemini and I have to create a read-only migration file and change its content so that ENUM can be seen as String. After that we change them from English to Chinese. Finally, we change them back to ENUM. Takes a while but it's doable.